pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillctl"
version = "0.1.0b1"
description = "Governance CLI for agent skills — validate, push, pull, and manage skill manifests"
readme = "README.md"
license = {text = "MPL-2.0"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
]
optimize = [
    "boto3>=1.34",
    "anthropic>=0.40",
]
dev = [
    "pytest>=7.0",
    "pytest-anyio>=0.0.0",
    "hypothesis>=6.0",
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]
all = [
    "skillctl[server,optimize]",
]

[project.scripts]
skillctl = "skillctl.cli:main"

[tool.setuptools.packages.find]
include = ["skillctl*"]
